pyproject.toml
[project]
name = "inkandquill-v2-poc"
version = "0.1.0"
description = "V2 Architecture POC - Docker-based migration from Azure"
authors = [{ name = "InkAndQuill Team", email = "contact@example.com" }]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "minio>=7.2.0",
    "sentence-transformers>=2.2.0",
    "openai>=1.0.0",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "numpy>=1.24.0",
    "psycopg2-binary>=2.9.0"
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"